modularized the animation API
